test deleting stuff
