+++ frontend/src/components/Chat/Chat_main.tsx
@@ -1,5 +1,5 @@
 import React, { useEffect, useState } from "react";
-import { fetchGetSet } from "src/ApiCalls/fetchers";
+import { fetchGetSet } from "src/functions/fetchers";
 import LeftBar from "src/components/Chat/LeftBar/LeftBar_main";
 import RightBar from "src/components/Chat/RightBar/RightBar_main";
 
@@ -29,7 +29,6 @@ function Chat() {
     }, [apiUrl_privateChats, apiUrl_publicChats]);
 
     return (
-        
         <div className="mainContainerColumn">
             <div className="h2">{selectedChat ? selectedChat.chatName : "Chat"}</div>
             <div className="chatMain">
