+++ frontend/src/components/Chat/RightBar/Members.tsx
@@ -1,5 +1,5 @@
 import React, { useEffect, useState } from "react";
-import { fetchGetSet } from "src/ApiCalls/fetchers";
+import { fetchGetSet } from "src/functions/fetchers";
 
 // DTO
 import { ChatListDTO, ParticipantListElementDTO } from "chat-dto";
@@ -11,18 +11,20 @@ interface membersProps {
 function Members(props: membersProps): React.JSX.Element {
     const [members, setMembers] = useState<ParticipantListElementDTO[]>([]);
     // temp to do
-    const userId = 0;
     const apiUrl =
-        process.env.REACT_APP_BACKEND_URL + "/chat/" + props.selectedChat.chatID + "/participants";
+        process.env.REACT_APP_BACKEND_URL +
+        "/chat/" +
+        props.selectedChat.chatID +
+        "/participants";
 
     useEffect(() => {
         fetchGetSet<ParticipantListElementDTO[]>(apiUrl, setMembers);
-    }, [props.selectedChat]);
+    }, [props.selectedChat, apiUrl]);
 
     function selectMember(userId) {
         console.log("selectMember: ", userId);
     }
-    
+
     /* 
    - interact with chat members
    - think of decision tree...
