+++ frontend/src/components/ManageProfile/FriendList.tsx
@@ -1,6 +1,6 @@
 import React, { useEffect, useState } from "react";
-import { fetchGetSet } from "src/ApiCalls/fetchers";
-import { blockUser, removeFriend } from "src/ApiCalls/userActions";
+import { fetchGetSet } from "src/functions/fetchers";
+import { blockUser, removeFriend } from "src/functions/userActions";
 
 // DTO
 import { IdAndNameDTO } from "user-dto";
