+++ frontend/src/components/ManageProfile/RequestInList.tsx
@@ -1,6 +1,6 @@
 import React, { useEffect, useState } from "react";
-import { fetchGetSet } from "src/ApiCalls/fetchers";
-import { acceptRequest, declineRequest } from "src/ApiCalls/userActions";
+import { fetchGetSet } from "src/functions/fetchers";
+import { acceptRequest, declineRequest } from "src/functions/userActions";
 
 // DTO
 import { IdAndNameDTO } from "user-dto";
