--- frontend/tsconfig.build.json
@@ -1,8 +1,0 @@
-{
-  "extends": "./tsconfig.json",
-  "exclude": ["node_modules", "test", "dist", "**/*spec.ts"],
-  "paths": {
-    "user-dto": ["../backend/src/user/user-dto"],
-    "chat-dto": ["../backend/src/chat/chat.DTOs"]
-  }
-}
