--- frontend/tsconfig.node.json
@@ -1,13 +1,0 @@
-{
-  "compilerOptions": {
-    "composite": true,
-    "skipLibCheck": true,
-    "module": "ESNext",
-    "moduleResolution": "bundler",
-    "allowSyntheticDefaultImports": true,
-    "paths": {
-      "user-dto": ["../backend/src/user/user-dto"]
-    }
-  },
-  "include": ["vite.config.ts"]
-}
